🔨 [#23] refactor: BodyType 엔티티 수정
체형 분석 관련 엔티티가 충돌이 나는 오류를 수정했습니다.
